New VariantEntityRepositoryDoubleFilter for maf, polyphen and sift to extend, has the getValue and getRelationalOperator methods
