+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/filter/VariantEntityRepositoryFilter.java
@@ -63,30 +63,6 @@ public abstract class VariantEntityRepositoryFilter<T> {
         return criteria;
     }
 
-    protected static Double getValueFromRelation(String relation) {
-        return Double.parseDouble(relation.replaceAll("[^\\d.]", ""));
-    }
-
-    protected static VariantEntityRepository.RelationalOperator getRelationalOperatorFromRelation(String relation) {
-        String relationalOperatorString = relation.replaceAll("[^<>=]", "");
-
-        switch (relationalOperatorString) {
-            case "=":
-                return VariantEntityRepository.RelationalOperator.EQ;
-            case ">":
-                return VariantEntityRepository.RelationalOperator.GT;
-            case "<":
-                return VariantEntityRepository.RelationalOperator.LT;
-            case ">=":
-                return VariantEntityRepository.RelationalOperator.GTE;
-            case "<=":
-                return VariantEntityRepository.RelationalOperator.LTE;
-            default:
-                throw new IllegalArgumentException();
-        }
-
-    }
-
     // equals and hashcode methods generated by intellij
     @Override
     public boolean equals(Object o) {
